# app/swagger.yml
openapi: '3.0.2'
info:
  title: Scoring App
  description: |-
    This is the documentation for the scoring app API endpoint. 
    Scoring app is an ML model built on scikit-learn framework. 
    The model receives one row of a pandas data frame and returns 
    the probabilities for the positive and negative class.
  version: '1.0'
  contact: 
    email: admin@somewhere.com
tags:
  - name: health
    description: Application health check
    
  - name: login
    description: User authentication
    
  - name: predict
    description: Predict scores

paths:
  /health:
    get:
      tags:
        - health
      summary: Health Check
      responses:
        '200':
          description: A JSON response indicating the health status of the application.
          content:
            /application.json:
              schema:
                type: object
                properties:
                  status:
                    type: string
                    description: The health status of the application.
   
  /login:
    post:
      tags:
        - login
      summary: User Authentication
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                username:
                  type: string
                  description: The username for authentication.
                password:
                  type: string
                  description: The password for authentication.
              required:
                - username
                - password
      responses:
        '200':
          description: A JSON response containing the access token.
          content:
            application/json:
              schema:
                type: object
                properties:
                  access_token:
                    type: string
                    description: The JWT access token.

  /predict:
    post:
      tags:
        - predict
      summary: Make Predictions
      description: |- 
        Enter a valid API key to access the API. 
        Enter the API key by clicking "Authorize" button on the top right corner of the screen.
        Add API key in this format [Bearer jwt_token]. Replace jwt_token with the 
        API key that you obtain when you log in. Keep "Bearer" in the beginning. 
        
        Example value: Bearer koasdkoajiajsfoa_737naksfjd
      parameters:
        - name: Authorization
          in: header
          required: false
          schema:
            type: string
            
      security:
        - JWT: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Predict'
            examples:
              example1:
                $ref: '#/components/schemas/Predict/examples/example1'
              example2:
                $ref: '#/components/schemas/Predict/examples/example2'
              
      responses:
        '200':
          description: A JSON response containing the prediction result.
          content:
            application/json:
              schema:
                type: object
                properties:
                  prediction:
                    type: array
                    description: The predicted probabilities

        '401':
          description: Unauthorized
          content:
            application/json:
              schema:
                type: object
                properties:
                  error:
                    type: string
                    description: The error message indicating unauthorized access.

        '400':
          description: Bad Request
          content:
            application/json:
              schema:
                type: object
                properties:
                  error:
                    type: string
                    description: The error message indicating bad request.

        '500':
          description: Internal Server Error
          content:
            application/json:
              schema:
                type: object
                properties:
                  error:
                    type: string
                    description: The error message indicating internal server error.
components:
  securitySchemes:
    JWT:
      type: apiKey
      in: header
      name: Authorization
  schemas:
    Predict:
      type: object
      properties:
        uuid:
          type: string
          description: The unique identifier
        default:
          type: number
          format: double
          description: The default value
        account_amount_added_12_24m:
          type: integer
          description: Account amount added in the last 12-24 months
        account_days_in_dc_12_24m:
          type: number
          format: double
          description: Account days in DC in the last 12-24 months
        account_days_in_rem_12_24m:
          type: number
          format: double
          description: Account days in rem in the last 12-24 months
        account_days_in_term_12_24m:
          type: number
          format: double
          description: Account days in term in the last 12-24 months
        account_incoming_debt_vs_paid_0_24m:
          type: number
          format: double
          description: Account incoming debt vs. paid in the last 0-24 months
        account_status:
          type: number
          format: double
          description: Account status
        account_worst_status_0_3m:
          type: number
          format: double
          description: Account worst status in the last 0-3 months
        account_worst_status_12_24m:
          type: number
          format: double
          description: Account worst status in the last 12-24 months
        account_worst_status_3_6m:
          type: number
          format: double
          description: Account worst status in the last 3-6 months
        account_worst_status_6_12m:
          type: number
          format: double
          description: Account worst status in the last 6-12 months
        age:
          type: integer
          description: The age
        avg_payment_span_0_12m:
          type: number
          format: double
          description: Average payment span in the last 0-12 months
        avg_payment_span_0_3m:
          type: number
          format: double
          description: Average payment span in the last 0-3 months
        merchant_category:
          type: string
          description: The merchant category
        merchant_group:
          type: string
          description: The merchant group
        has_paid:
          type: boolean
          description: Indicates whether the payment has been made
        max_paid_inv_0_12m:
          type: number
          format: double
          description: Maximum paid invoice in the last 0-12 months
        max_paid_inv_0_24m:
          type: number
          format: double
          description: Maximum paid invoice in the last 0-24 months
        name_in_email:
          type: string
          description: The name in the email
        num_active_div_by_paid_inv_0_12m:
          type: number
          format: double
          description: Number of active div by paid invoice in the last 0-12 months
        num_active_inv:
          type: integer
          description: Number of active invoices
        num_arch_dc_0_12m:
          type: integer
          description: Number of archived DC in the last 0-12 months
        num_arch_dc_12_24m:
          type: integer
          description: Number of archived DC in the last 12-24 months
        num_arch_ok_0_12m:
          type: integer
          description: Number of archived OK in the last 0-12 months
        num_arch_ok_12_24m:
          type: integer
          description: Number of archived OK in the last 12-24 months
        num_arch_rem_0_12m:
          type: integer
          description: Number of archived REM in the last 0-12 months
        num_arch_written_off_0_12m:
          type: number
          format: double
          description: Number of archived written off in the last 0-12 months
        num_arch_written_off_12_24m:
          type: number
          format: double
          description: Number of archived written off in the last 12-24 months
        num_unpaid_bills:
          type: integer
          description: Number of unpaid bills
        status_last_archived_0_24m:
          type: integer
          description: Status last archived in the last 0-24 months
        status_2nd_last_archived_0_24m:
          type: integer
          description: Status 2nd last archived in the last 0-24 months
        status_3rd_last_archived_0_24m:
          type: integer
          description: Status 3rd last archived in the last 0-24 months
        status_max_archived_0_6_months:
          type: integer
          description: Status max archived in the last 0-6 months
        status_max_archived_0_12_months:
          type: integer
          description: Status max archived in the last 0-12 months
        status_max_archived_0_24_months:
          type: integer
          description: Status max archived in the last 0-24 months
        recovery_debt:
          type: integer
          description: Recovery debt
        sum_capital_paid_account_0_12m:
          type: integer
          description: Sum capital paid account in the last 0-12 months
        sum_capital_paid_account_12_24m:
          type: integer
          description: Sum capital paid account in the last 12-24 months
        sum_paid_inv_0_12m:
          type: integer
          description: Sum paid invoice in the last 0-12 months
        time_hours:
          type: number
          format: double
          description: Time hours
        worst_status_active_inv:
          type: number
          format: double
          description: Worst status active invoice
      examples:
        example1:
          value:
            uuid: "123abc"
            default: 10.5
            account_amount_added_12_24m: 100
            account_days_in_dc_12_24m: 10.5
            account_days_in_rem_12_24m: 20.5
            account_days_in_term_12_24m: 15.0
            account_incoming_debt_vs_paid_0_24m: 0.8
            account_status: 1.0
            account_worst_status_0_3m: 0.5
            account_worst_status_12_24m: 0.2
            account_worst_status_3_6m: 0.3
            account_worst_status_6_12m: 0.4
            age: 35
            avg_payment_span_0_12m: 5.0
            avg_payment_span_0_3m: 3.0
            merchant_category: "Retail"
            merchant_group: "Group A"
            has_paid: true
            max_paid_inv_0_12m: 150.0
            max_paid_inv_0_24m: 200.0
            name_in_email: "John Doe"
            num_active_div_by_paid_inv_0_12m: 2.0
            num_active_inv: 3
            num_arch_dc_0_12m: 1
            num_arch_dc_12_24m: 2
            num_arch_ok_0_12m: 3
            num_arch_ok_12_24m: 4
            num_arch_rem_0_12m: 5
            num_arch_written_off_0_12m: 0.5
            num_arch_written_off_12_24m: 1.0
            num_unpaid_bills: 2
            status_last_archived_0_24m: 1
            status_2nd_last_archived_0_24m: 2
            status_3rd_last_archived_0_24m: 3
            status_max_archived_0_6_months: 1
            status_max_archived_0_12_months: 2
            status_max_archived_0_24_months: 3
            recovery_debt: 200
            sum_capital_paid_account_0_12m: 300
            sum_capital_paid_account_12_24m: 400
            sum_paid_inv_0_12m: 500
            time_hours: 8.0
            worst_status_active_inv: 0.1
        example2:
          value:
            uuid: "456def"
            default: 15.0
            account_amount_added_12_24m: 200
            account_days_in_dc_12_24m: 15.0
            account_days_in_rem_12_24m: 25.0
            account_days_in_term_12_24m: 20.0
            account_incoming_debt_vs_paid_0_24m: 0.7
            account_status: 0.8
            account_worst_status_0_3m: 0.4
            account_worst_status_12_24m: 0.3
            account_worst_status_3_6m: 0.2
            account_worst_status_6_12m: 0.1
            age: 40
            avg_payment_span_0_12m: 6.0
            avg_payment_span_0_3m: 4.0
            merchant_category: "Technology"
            merchant_group: "Group B"
            has_paid: false
            max_paid_inv_0_12m: 180.0
            max_paid_inv_0_24m: 250.0
            name_in_email: "Jane Doe"
            num_active_div_by_paid_inv_0_12m: 2.5
            num_active_inv: 4
            num_arch_dc_0_12m: 2
            num_arch_dc_12_24m: 3
            num_arch_ok_0_12m: 4
            num_arch_ok_12_24m: 5
            num_arch_rem_0_12m: 6
            num_arch_written_off_0_12m: 0.7
            num_arch_written_off_12_24m: 1.5
            num_unpaid_bills: 3
            status_last_archived_0_24m: 2
            status_2nd_last_archived_0_24m: 3
            status_3rd_last_archived_0_24m: 4
            status_max_archived_0_6_months: 2
            status_max_archived_0_12_months: 3
            status_max_archived_0_24_months: 4
            recovery_debt: 250
            sum_capital_paid_account_0_12m: 350
            sum_capital_paid_account_12_24m: 450
            sum_paid_inv_0_12m: 550
            time_hours: 10.0
            worst_status_active_inv: 0.2